+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog][Keep a Changelog] and this project adheres to [Semantic Versioning][Semantic Versioning].
 
 ## [Unreleased]
+### Added
+- `.assertWhen(condition, test, failureMessage)` to only run an assertion conditionally.
 
 ## [0.1.0] - 2020-04-16
 ### Added